requirements.txt
accelerator==2024.9.13
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
asyncer==0.0.7
attrs==24.2.0
av==12.3.0
bidict==0.23.1
bottle==0.12.25
certifi==2024.8.30
cffi==1.17.1
chainlit==1.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dataclasses-json==0.6.7
datasets==3.0.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dnspython==2.0.0
ecdsa==0.19.0
fastapi==0.112.4
fastapi-socketio==0.0.10
filelock==3.16.1
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.6.1
google-crc32c==1.6.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.25.1
hyperframe==6.0.1
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.4.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.1
langchain-openai==0.2.0
langchain-qdrant==0.1.4
langchain-text-splitters==0.3.0
langgraph==0.2.35
langgraph-checkpoint==2.0.0
langsmith==0.1.121
Lazify==0.4.0
MarkupSafe==3.0.1
marshmallow==3.22.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.1
numpy==1.26.4
openai==1.51.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.10.7
packaging==23.2
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
portalocker==2.10.1
protobuf==4.25.5
pyarrow==17.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyee==12.0.0
PyJWT==2.9.0
pylibsrtp==0.10.0
pymongo==3.11.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyOpenSSL==24.2.1
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-graphql-client==0.4.3
python-jose==3.3.0
python-multipart==0.0.9
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.11.2
ragas==0.2.0b0
regex==2024.9.11
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.0
setproctitle==1.3.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.37.2
sympy==1.13.3
syncer==2.0.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.1
tomli==2.0.1
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uptrace==1.26.0
urllib3==2.2.3
uvicorn==0.25.0
waitress==3.0.0
watchfiles==0.20.0
websockets==13.1
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.13.1
zipp==3.20.2
